+++ src/main/java/joshie/harvest/api/cooking/IFridge.java
@@ -1,0 +1,7 @@
+package joshie.harvest.api.cooking;
+
+import net.minecraft.inventory.IInventory;
+
+public interface IFridge {
+    IInventory getContents();
+}