+++ src/main/java/joshie/harvest/cooking/tile/FridgeData.java
@@ -117,6 +117,7 @@ public class FridgeData implements IInventory {
 
         player.worldObj.addBlockEvent(tile.getPos(), tile.getBlockType(), 1, players);
         tile.getWorld().playSound(null, tile.getPos().getX(), tile.getPos().getY(), tile.getPos().getZ(), HFSounds.FRIDGE, SoundCategory.BLOCKS, 2F, tile.getWorld().rand.nextFloat() * 0.1F + 0.9F);
+        tile.saveAndRefresh(); //Update the client about this
     }
 
     @Override