+++ src/main/java/joshie/harvest/npcs/gift/GiftsAshlee.java
@@ -13,7 +13,6 @@ public class GiftsAshlee extends Gifts {
         stackRegistry.register(Ore.of("cropBanana"), Quality.AWESOME);
         stackRegistry.register(Ore.of("cropCorn"), Quality.AWESOME);
         stackRegistry.register(MEAL.getStackFromEnum(Meal.CORN_BAKED), Quality.AWESOME);
-        categoryRegistry.put(HERB, Quality.GOOD);
         categoryRegistry.put(EGG, Quality.GOOD);
         categoryRegistry.put(FRUIT, Quality.GOOD);
         stackRegistry.register(MEAL.getStackFromEnum(Meal.KETCHUP), Quality.BAD);