+++ src/main/java/joshie/harvest/npcs/gift/GiftsJade.java
@@ -1,6 +1,8 @@
 package joshie.harvest.npcs.gift;
 
 import joshie.harvest.api.core.Ore;
+import joshie.harvest.gathering.HFGathering;
+import joshie.harvest.gathering.block.BlockNature.NaturalBlock;
 import net.minecraft.init.Blocks;
 import net.minecraft.init.Items;
 import net.minecraft.item.ItemStack;
@@ -12,6 +14,7 @@ public class GiftsJade extends Gifts {
     public GiftsJade() {
         stackRegistry.register(Blocks.YELLOW_FLOWER, Quality.AWESOME);
         stackRegistry.register(new ItemStack(Blocks.RED_FLOWER, 1, 2), Quality.AWESOME);
+        stackRegistry.register(HFGathering.NATURE.getStackFromEnum(NaturalBlock.LAVENDER), Quality.AWESOME);
         stackRegistry.register(Items.RABBIT_FOOT, Quality.GOOD);
         categoryRegistry.put(FLOWER, Quality.GOOD);
         categoryRegistry.put(PLANT, Quality.GOOD);